requirements.txt
streamlit
python-pptx
transformers
torch
sentencepiece
sacremoses
tesseract-ocr
libtesseract-dev
Pillow
pytesseract